tml-poc-try1/my_agent/requirements_kdb_v3.txt
strands-agents
strands-agents-tools
strands-agents-builder
uv
chromadb>=0.5
sentence-transformers>=2.6
pypdf
